Working chat filtering, formatting, etc.
I was working on this project months ago and somewhat forgot about it as
I have become quite busy with other things. I intended to finish it
before I committed it, but I realize that it's best to push what I have
now rather than later. There's still a lot to do with the plugin, but so
far it works pretty well. Enjoy.